+++ SBBMobileDesignSystemSwiftUIDemo/UI/DetailViews/PaginationViewDemo.swift
@@ -37,7 +37,7 @@ struct PaginationViewDemo: View {
         }
             .padding(16)
             .navigationBarTitle("Pagination")
-            .background(SBBColor.background)
+            .background(SBBColor.background.edgesIgnoringSafeArea(.bottom))
             .colorScheme(colorScheme)
     }
 }