+++ SBBMobileDesignSystemSwiftUIDemo/UI/DetailViews/TypographyView/TypographyView.swift
@@ -21,13 +21,13 @@ struct TypographyView: View {
             ScrollView(showsIndicators: false) {
                 VStack(spacing: 16) {
                     if fontView == 1 {
-                        SBBInfoView(title: Text("ReadMe   -   .font"), detail: Text("Using the .font() View Modifier, you can select all available SBB Fonts. However this does not apply SBB specific LineSpacing. If you want SBB specific LineSpacing, use the .sbbFont() View Modifier instead."), expanded: $infoViewExpanded)
+                        SBBInfoViewCollapsible(title: Text("ReadMe   -   .font"), detail: Text("Using the .font() View Modifier, you can select all available SBB Fonts. However this does not apply SBB specific LineSpacing. If you want SBB specific LineSpacing, use the .sbbFont() View Modifier instead."), expanded: $infoViewExpanded)
                         TypographyFontView()
                     } else if fontView == 2 {
-                        SBBInfoView(title: Text("ReadMe   -   .sbbFont"), detail: Text("Using the .sbbFont() View Modifier, you can select all available SBB Fonts. SBB specific LineSpacing is also applied."), expanded: $infoViewExpanded)
+                        SBBInfoViewCollapsible(title: Text("ReadMe   -   .sbbFont"), detail: Text("Using the .sbbFont() View Modifier, you can select all available SBB Fonts. SBB specific LineSpacing is also applied."), expanded: $infoViewExpanded)
                         TypographySBBFontView()
                     } else if fontView == 3 {
-                        SBBInfoView(title: Text("ReadMe   -   Custom Font"), detail: Text("Using the .font(.sbbFont(size: 10)) you can create your own Font based on the available SBB Font Styles by specifying a size of your choice"), expanded: $infoViewExpanded)
+                        SBBInfoViewCollapsible(title: Text("ReadMe   -   Custom Font"), detail: Text("Using the .font(.sbbFont(size: 10)) you can create your own Font based on the available SBB Font Styles by specifying a size of your choice"), expanded: $infoViewExpanded)
                         TypographyCustomFontView()
                     }
                 }
@@ -35,7 +35,7 @@ struct TypographyView: View {
         }
             .navigationBarTitle("Typography")
             .padding()
-            .background(SBBColor.background)
+            .background(SBBColor.background.edgesIgnoringSafeArea(.bottom))
             .colorScheme(colorScheme)
             
     }